Provide `draw` and `solve` CLI commands
Implement CLI interface and commands.

The `draw` command renders the internal state graph in graphviz DOT
format.

The `solve` command reduces the state graph to its LLB result and
outputs the definition as protobuf. This can be piped to `buildctl
build`. Although this command is useful in the proof of concept, it will
likely be limited to a debugging interface in the final form, the reason
being that only a single image result can produced in this way, and one
goal of the project is to provide manifest indexes as the canonical
output form.