compiler: Support a `copy` top-level state operation
The new `#CopyOp` definition implements a more user friendly top-level
`copy` operation that can copy multiple sources in a single instruction,
unlike the `file.copy` file action which can only copy a single file or
directory at a time.

Note that `#CopyOp` is a pure-CUE implementation that utilizes the `ops`
aggregate in combination with `file`.