schema: Support referencing chains using actual CUE references
Current performance issues in the CUE evaluator(s) prevent the use of
actual CUE references in composition such as `{ copy: ".", from:
chains.foo }`. As a workaround, the schema has been relying on string
based references that are resolved at CUE-to-LLB compile time.

Refactor the schema and compiler to support these direct references
anyhow, and use the existing `config/load_test.go` as a benchmark of
sorts to keep the performance issue surfaced. This change will live on
an experimental branch for now, and be merged once this test can pass in
a reasonable amount of time.