config: Benchmark `#ChainStringRef` vs `#Chain` as `#ChainRef`
The `#ChainRef: #ChainStringRef | #Chain` definition allows users to
make use of direct CUE references in their configuration.

    chains: {
      project: mainContext: true
      modules: [
        { copy: ["go.mod", "go.sum"], from: project }
        // ...
      ]
    }

However, it seems there are still substantial performance differences
when doing this vs. using the `#ChainStringRef` form.

    chains: {
      project: mainContext: true
      modules: [
        { copy: ["go.mod", "go.sum"], from: "project" }
        // ...
      ]
    }

Add two benchmarks to compare performance between `#Chain` and
`#ChainStringRef` so we can know when to start promoting use of the
former as the CUE's evaluator improves.